+++ src/task3/Study.java
@@ -13,10 +13,4 @@ public class Study {
     }
 }
 
-class JavaProgram {
-    public static void main(String[] args) {
-        Study study = new Study("Изучение Java - это просто!");
-        System.out.println(study.printCourse());;
-    }
 
-}
